Revert payment page to viewcart popup
