+++ frontend/src/Components/ViewCart/ViewCartPopup.css
@@ -1,0 +1,47 @@
+.cart-table {
+    width: 100%;
+    border-collapse: collapse;
+  }
+  
+    .cart-table td {
+    text-align: right;
+  }
+  .cart-table th {
+    background-color: #f2f2f2;
+    border: 1px solid #ccc;
+    padding: 5px;
+    text-align: left;
+  }
+  
+  
+  .cart-table td.total-price {
+
+    color: blue;
+  }
+
+  .total-price-cell {
+    color: red;
+  }
+
+  .cart-table tfoot tr {
+    border: none; /* Remove border for tfoot */
+  }
+  
+  .cart-table tfoot td {
+    border: none; /* Remove border for tfoot cells */
+  }
+  
+  .quantity-input {
+    width: 50px; /* Adjust the width as needed */
+  }
+  
+  .button-container {
+    display: flex;
+    justify-content: space-between;
+    margin-top: 20px;
+  }
+  
+  .button-container button {
+    padding: 10px 20px;
+  }
+
