--- frontend/src/Containers/PaymentPage/Payment.css
@@ -1,64 +1,0 @@
-.payment-page {
-  background-color: white;
-  max-width: 800px; /* Adjust the max-width as needed */
-  width: 100%;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px;
-}
-
-.payment-content {
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
-.cart-table {
-  width: 100%;
-  border-collapse: collapse;
-}
-
-  .cart-table td {
-  text-align: right;
-}
-.cart-table th {
-  background-color: #f2f2f2;
-  border: 1px solid #ccc;
-  padding: 5px;
-  text-align: left;
-}
-
-
-.cart-table td.total-price {
-
-  color: blue;
-}
-
-.total-price-cell {
-  color: red;
-}
-
-.cart-table tfoot tr {
-  border: none; /* Remove border for tfoot */
-}
-
-.cart-table tfoot td {
-  border: none; /* Remove border for tfoot cells */
-}
-
-.quantity-input {
-  width: 50px; /* Adjust the width as needed */
-}
-
-.button-container {
-  display: flex;
-  justify-content: space-between;
-  margin-top: 20px;
-  width: 100%;
-}
-
-.button-container button {
-  padding: 10px 20px;
-}
